redefine exterior normal with explicit reference
The sign of the exterior normal follows the sign of the determinant of the
rootgradient. For the rectilinear topologies this is usually fine. However, for
the simplex topology the sign typically flips from one element to another and
the only way to change this is to introduce a negative scaling in the transform
chain for selected elements. This is exactly how `mesh.unitsquare` solves this
problem.

This patch eliminates the dependency on the sign of the determinant of
transform chains. The `exterior` argument of `function.normal` must be a
reference geometry (with the same dimension as the topology). The rootgradient
is replaced by a gradient w.r.t. the reference.