add transform.Point
This patch adds an implementation of `TransformItem` for points (`fromdims=0`).
This is seemingly useless, as the implementation doesn't add anything to its
base `Matrix`. However, the constructor is simpler than `Matrix` and `Matrix`
is currently used for points instances only. Replacing thoses instances with
`Point` is done in a follow-up patch.