remove MKL_THREADING_LAYER override
MS Windows is reported to have issues with MKL's TBB threading layer. Since
this value is set in _mkl.py, this means scripts will fail on Windows with
OSError [WinError] unless they use cli.run (which changes the threading layer
to sequential). To fix scripts that do not make use of cli.run, this patch
remove the override of the MKL_THREADING_LAYER variable.